fix: tests and associated bugs
This is a placeholder which contains all the fixes and improvements
needed to get all the existing tests running error free.

There are a small number of tests which have been marked as skipped
due to flaky behaviour or significant functionality needed to make
them work.

The intention is to split this out into many more consumable changes.

This should not be merged as is.